pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-devops-agent"
version = "1.0.0"
description = "Production-ready AWS DevOps automation agent using Strands + Bedrock Agent Core + MCP"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Caylent", email = "info@caylent.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "strands-agents",
    "bedrock-agentcore",
    "boto3",
    "pandas",
    "openpyxl",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
    "pre-commit",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "myst-parser",
]

[project.scripts]
aws-devops-agent = "aws_devops_agent.main:main"

[project.urls]
Homepage = "https://github.com/caylent/aws-devops-agent"
Documentation = "https://github.com/caylent/aws-devops-agent/docs"
Repository = "https://github.com/caylent/aws-devops-agent.git"
Issues = "https://github.com/caylent/aws-devops-agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
]
